integrated sc activity and activity kernel

+++ src/kernels.cu
@@ -122,7 +122,7 @@ __global__ void updatePFPCOutGPU(uint32_t *apBuf, uint32_t *delay,
 //**---------------SC Kernels-------------------**
 
 __global__ void calcSCActivity(float *vSC, float *gPFSC, float *threshSC, uint8_t *apSC,
-	uint32_t *apBufSC,  float *gInputSumPFSC, float eLeakSC, float gLeakSC, float gIncPFSC,
+	uint32_t *apBufSC, uint32_t *gInputSumPFSC, float eLeakSC, float gLeakSC, float gIncPFSC,
 	float gDecPFSC, float threshRestSC, float threshMaxSC, float threshDecaySC)
 {
 	int i = blockIdx.x*blockDim.x+threadIdx.x;
@@ -442,7 +442,7 @@ void callPCActKernel(cudaStream_t &st, unsigned int numBlocks, unsigned int numP
 }
 
 void callSCActKernel(cudaStream_t &st, unsigned int numBlocks, unsigned int numSCPerBlock, 
-	float *vSC, float *gPFSC, float *threshSC, uint8_t *apSC, uint32_t *apBufSC, float *gInputSumPFSC,
+	float *vSC, float *gPFSC, float *threshSC, uint8_t *apSC, uint32_t *apBufSC, uint32_t *gInputSumPFSC,
 	float eLeakSC, float gLeakSC, float gIncPFSC, float gDecPFSC, float threshRestSC, float threshMaxSC,
 	float threshDecaySC)
 {
