+++ src/kernels.h
@@ -29,7 +29,7 @@ void callPCActKernel(cudaStream_t &st, unsigned int numBlocks, unsigned int numP
 	float threshRestPC, float threshMaxPC, float threshDecayPC);
 
 void callSCActKernel(cudaStream_t &st, unsigned int numBlocks, unsigned int numSCPerBlock, 
-	float *vSC, float *gPFSC, float *threshSC, uint8_t *apSC, uint32_t *apBufSC, float *gInputSumPFSC,
+	float *vSC, float *gPFSC, float *threshSC, uint8_t *apSC, uint32_t *apBufSC, uint32_t *gInputSumPFSC,
 	float eLeakSC, float gLeakSC, float gIncPFSC, float gDecPFSC, float threshRestSC, float threshMaxSC,
 	float threshDecaySC);
